FSD構成のpages配下にapp/page.tsxの中身を移動
